pyproject.toml
[project]
name = "chatwithpdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.7",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.30",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.13",
    "pypdf>=6.10.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.44.0",
]
